+++ Tests/Unit/Mapper/Fixtures/TestingMapper.php
@@ -12,8 +12,7 @@ use OliverKlee\Oelib\Tests\Unit\Model\Fixtures\TestingModel;
 class TestingMapper extends \Tx_Oelib_DataMapper
 {
     /**
-     * @var string the name of the database table for this mapper,
-     *             must not be empty in subclasses
+     * @var string the name of the database table for this mapper
      */
     protected $tableName = 'tx_oelib_test';
 
@@ -21,4 +20,17 @@ class TestingMapper extends \Tx_Oelib_DataMapper
      * @var string the model class name for this mapper, must not be empty
      */
     protected $modelClassName = TestingModel::class;
+
+    /**
+     * @var string[] the (possible) relations of the created models in the format DB column name => mapper name
+     */
+    protected $relations = [
+        'friend' => TestingMapper::class,
+        'owner' => \Tx_Oelib_Mapper_FrontEndUser::class,
+        'children' => TestingMapper::class,
+        'related_records' => TestingMapper::class,
+        'composition' => \Tx_Oelib_Tests_LegacyUnit_Fixtures_TestingChildMapper::class,
+        'composition2' => \Tx_Oelib_Tests_LegacyUnit_Fixtures_TestingChildMapper::class,
+        'bidirectional' => TestingMapper::class,
+    ];
 }