+++ Tests/Unit/Model/Fixtures/TestingModel.php
@@ -88,4 +88,66 @@ class TestingModel extends \Tx_Oelib_Model
     {
         parent::setLoadStatus($status);
     }
+
+    /**
+     * Gets the "friend" data item. This is an n:1 relation.
+     *
+     * @return TestingModel
+     */
+    public function getFriend()
+    {
+        return $this->getAsModel('friend');
+    }
+
+    /**
+     * Gets the "children" data item. This is a 1:n relation.
+     *
+     * @return \Tx_Oelib_List<TestingModel>
+     */
+    public function getChildren()
+    {
+        return $this->getAsList('children');
+    }
+
+    /**
+     * Gets the "related_records" data item. This is an m:n relation.
+     *
+     * @return \Tx_Oelib_List<TestingModel>
+     */
+    public function getRelatedRecords()
+    {
+        return $this->getAsList('related_records');
+    }
+
+    /**
+     * Adds a related record.
+     *
+     * @param TestingModel $record
+     *
+     * @return void
+     */
+    public function addRelatedRecord(TestingModel $record)
+    {
+        $this->getRelatedRecords()->add($record);
+    }
+
+    /**
+     * Gets the "bidirectional" data item. This is an m:n relation.
+     *
+     * @return \Tx_Oelib_List<TestingModel>
+     */
+    public function getBidirectional()
+    {
+        return $this->getAsList('bidirectional');
+    }
+
+    /**
+     * Gets the "composition" data item. This is an 1:n relation.
+     *
+     * @return \Tx_Oelib_List<\Tx_Oelib_Tests_LegacyUnit_Fixtures_TestingChildModel>
+     */
+    public function getComposition()
+    {
+        return $this->getAsList('composition');
+    }
 }